requirements.txt
djangorestframework==3.15.1
Django==4.2.11
uvicorn==0.29.0
django-filter==24.2
drf-spectacular==0.27.2
coreapi==2.3.3
fire==0.5.0